chore: extract lib fn
