+++ src/app/posts/[...slug]/page.tsx
@@ -1,8 +1,8 @@
 import BackButton from "@/components/back-button";
 import Metadata from "@/components/metadata";
 import WebmentionsList from "@/components/webmentions.tsx";
+import { getContentAsHTML } from "@/lib/get-content-as-html";
 import { getPostBySlug } from "@/lib/get_post_by_slug";
-import markdownToHtml from "@/lib/markdown_to_html";
 
 type Params = {
   params: { slug: string[] };
@@ -19,10 +19,6 @@ type PostType = {
   };
   content: string;
 };
-
-export async function getContentAsHTML(content: string) {
-  return await markdownToHtml(content);
-}
 
 export default async function Page({ params, searchParams }: Params) {
   const post = getPostBySlug(params.slug.join("/"), [
