+++ src/lib/get-content-as-html.ts
@@ -1,0 +1,5 @@
+import markdownToHtml from "./markdown_to_html";
+
+export async function getContentAsHTML(content: string) {
+  return await markdownToHtml(content);
+}
